+++ mne_icalabel/annotation/__init__.py
@@ -1,0 +1,1 @@
+from .bids import mark_component, write_channels_tsv